+++ packages/components/src/Autocomplete/Autocomplete.tsx
@@ -12,7 +12,15 @@ type OptionCollection = XOR<Option[], GroupOption[]>;
 interface AutocompleteProps
   extends Pick<
     FormFieldProps,
-    "size" | "onBlur" | "onFocus" | "invalid" | "name" | "validations"
+    | "inputRef"
+    | "invalid"
+    | "name"
+    | "onBlur"
+    | "onFocus"
+    | "prefix"
+    | "size"
+    | "suffix"
+    | "validations"
   > {
   /**
    * Initial options to show when user first focuses the Autocomplete
@@ -59,25 +67,21 @@ interface AutocompleteProps
   readonly placeholder: string;
 }
 
-/**
- * Max statements disabled here to make room for the
- * debounce functions.
- */
-// eslint-disable-next-line max-statements
+// Max statements increased to make room for the debounce functions
+/* eslint max-statements: ["error", 14] */
 export function Autocomplete({
   initialOptions = [],
   value,
   allowFreeForm = true,
   size = undefined,
-  invalid,
   debounce: debounceRate = 300,
   onChange,
   getOptions,
   placeholder,
   onBlur,
   onFocus,
-  name,
   validations,
+  ...inputProps
 }: AutocompleteProps) {
   const [options, setOptions] = useState(initialOptions);
   const [menuVisible, setMenuVisible] = useState(false);
@@ -100,14 +104,13 @@ export function Autocomplete({
       <InputText
         autocomplete={false}
         size={size}
-        invalid={invalid}
         value={inputText}
         onChange={handleInputChange}
         placeholder={placeholder}
         onFocus={handleInputFocus}
         onBlur={handleInputBlur}
-        name={name}
         validations={validations}
+        {...inputProps}
       />
       {menuVisible && (
         <Menu