+++ packages/components/CHANGELOG.md
@@ -3,6 +3,116 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.24.1](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.24.0...@jobber/components@4.24.1) (2023-08-03)
+
+**Note:** Version bump only for package @jobber/components
+
+
+
+
+
+# [4.24.0](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.23.66...@jobber/components@4.24.0) (2023-08-02)
+
+
+### Features
+
+* **components:** InputText Props for Autocomplete ([#1465](https://github.com/GetJobber/atlantis/issues/1465)) ([7a978ef](https://github.com/GetJobber/atlantis/commit/7a978ef8185d70a57a161e61ea795b1c7b6fd429))
+
+
+
+
+
+## [4.23.66](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.23.65...@jobber/components@4.23.66) (2023-08-02)
+
+**Note:** Version bump only for package @jobber/components
+
+
+
+
+
+## [4.23.65](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.23.64...@jobber/components@4.23.65) (2023-08-02)
+
+**Note:** Version bump only for package @jobber/components
+
+
+
+
+
+## [4.23.64](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.23.63...@jobber/components@4.23.64) (2023-08-02)
+
+**Note:** Version bump only for package @jobber/components
+
+
+
+
+
+## [4.23.63](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.23.62...@jobber/components@4.23.63) (2023-08-01)
+
+**Note:** Version bump only for package @jobber/components
+
+
+
+
+
+## [4.23.62](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.23.61...@jobber/components@4.23.62) (2023-08-01)
+
+
+### Bug Fixes
+
+* update FormatTime logic for 12 or 24 hours ([#1442](https://github.com/GetJobber/atlantis/issues/1442)) ([67869ce](https://github.com/GetJobber/atlantis/commit/67869cea1ceb70a2e56cc5cee73e6fe26907edb1))
+
+
+
+
+
+## [4.23.61](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.23.60...@jobber/components@4.23.61) (2023-08-01)
+
+**Note:** Version bump only for package @jobber/components
+
+
+
+
+
+## [4.23.60](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.23.59...@jobber/components@4.23.60) (2023-08-01)
+
+**Note:** Version bump only for package @jobber/components
+
+
+
+
+
+## [4.23.59](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.23.58...@jobber/components@4.23.59) (2023-07-31)
+
+**Note:** Version bump only for package @jobber/components
+
+
+
+
+
+## [4.23.58](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.23.57...@jobber/components@4.23.58) (2023-07-31)
+
+**Note:** Version bump only for package @jobber/components
+
+
+
+
+
+## [4.23.57](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.23.56...@jobber/components@4.23.57) (2023-07-31)
+
+**Note:** Version bump only for package @jobber/components
+
+
+
+
+
+## [4.23.56](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.23.55...@jobber/components@4.23.56) (2023-07-31)
+
+**Note:** Version bump only for package @jobber/components
+
+
+
+
+
 ## [4.23.55](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.23.54...@jobber/components@4.23.55) (2023-07-31)
 
 **Note:** Version bump only for package @jobber/components